[FIXED HUDSON-2260] Turns out the test failure I saw during a release was a real problem, not a timing issue.
I guess the good news is that the tests are now indeed catching this kind of problems, which is a big step forward.

The sad part is that I ignored it and went ahead with a release anyway.

This bug is fixed in 1.251.

git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@11693 71c3de6d-444a-0410-be80-ed276b4c234a

Originally-Committed-As: 02093d1d61adc725ca61144d892d2b4d22ea82f4