Set the compatibleSinceVersion flag
Version 1.16 introduced a change (JENKINS-21475) in the configuration file which is not backward compatible in terms of plugin downgrades. Although plugin downgrades are not usually supported and not granted to work, additions to the config file are just ignored and old features still work, but not in this case, because there is a structure change in the file which makes older plugin versions unable to know what to load from disk.

The compat warning message says: `Warning: [...] and/or you may not be able to cleanly revert to the prior version without manually restoring old settings [...]`. So I think changes for JENKINS-21475 justify the need of the `compatibleSinceVersion` flag here.