[JENKINS-65117] check a parsed DN not a String
A DN is hierachical and contains attribute names and their values.
an attribute name is case insensitive so when comparing a String
representaion of a DN you will have incorrect results if one side has a
different case for the attribute name than the other.

The solution is to parse the DN and compare the parsed results.

Manually tested with the planet express test DB and with a user that has
some characters that need to be escaped as part of its CN (a comma).